feat: migrate from VideoService to MediaExportService and remove FFmpeg dependencies
